removing label
