+++ packages/libs/react-ui/src/components/Form/FormFieldWrapper/FormFieldHeader/FormFieldHeader.css.ts
@@ -1,6 +1,15 @@
 import { sprinkles } from '@theme/sprinkles.css';
 import { style } from '@vanilla-extract/css';
 
+export const labelClass = style([
+  sprinkles({
+    fontSize: '$sm',
+    color: '$foreground',
+    fontWeight: '$bold',
+    textTransform: 'capitalize',
+  }),
+]);
+
 export const headerClass = style([
   sprinkles({
     display: 'flex',
