+++ packages/libs/react-ui/src/components/Typography/index.ts
@@ -3,5 +3,4 @@ export {
   type IGradientTextProps,
 } from './GradientText/GradientText';
 export { Heading, type IHeadingProps } from './Heading/Heading';
-export { Label, type ILabelProps } from './Label/Label';
 export { Text, type ITextProps } from './Text/Text';
