+++ packages/libs/react-ui/src/index.ts
@@ -25,7 +25,6 @@ export type {
   IIconButtonProps,
   IIconProps,
   IInputProps,
-  ILabelProps,
   ILabelValue,
   ILinkProps,
   IMaskedValueProps,
@@ -91,7 +90,6 @@ export {
   IconButton,
   Input,
   InputCopy,
-  Label,
   Link,
   MaskedValue,
   NavFooter,
